+++ src/steps/trackReplaceStep.ts
@@ -29,7 +29,6 @@ import { ExposedReplaceStep, ExposedSlice } from '../types/pm'
 import { ChangeStep } from '../types/step'
 import { TrTrackingContext } from '../types/track'
 import { log } from '../utils/logger'
-import * as trackUtils from '../utils/track-utils'
 import { mapChangeSteps } from '../utils/mapChangeStep'
 import { NewEmptyAttrs } from '../attributes/types'
 import { isStructuralChange } from '../changes/qualifiers'
