+++ src/track-changes/fixAndHandleSelection.ts
