+++ src/track-changes/processStepsBeforeTracking.ts
